wpa_supplicant: Allow building on ANDROID if no private library is defined
If there is no private library defined on Android, the build will
fail due to missing reference to wpa_driver_nl80211_driver_cmd.

This patch modifies the makefile to introduce a ANDROID_PLIB_STUB
definition, which is set if BOARD_WPA_SUPPLICANT_PRIVATE_LIB
is not set and allows us to avoid
referencing wpa_driver_nl80211_driver_cmd in that case.

Change-Id: Icc7339286f312088ef956ff83b7ec47539023740
Signed-off-by: John Stultz <john.stultz@linaro.org>